Add libFuzzer infrastructure and fix address parsing DNS timeout
Added comprehensive fuzzing infrastructure:
- Created 4 fuzzing targets: SAN validation, PEM parsing, certificate
  validation, and address parsing
- Added MTLS_ENABLE_FUZZING CMake option with per-target compilation flags
- Created reusable fuzzing utilities and corpus seeds
- All fuzzers use AddressSanitizer for memory error detection

Fixed PERF-001 performance issue:
- Fuzzing discovered that .local TLD triggered 25-second mDNS timeouts
- Added validate_hostname() to reject control characters and problematic TLDs
- Hostname validation now runs before expensive getaddrinfo() calls
- Performance improvement: 15x faster execution (94 to 1,402 exec/s)

This prevents DNS-based DoS attacks where attackers control address input.

Also fixed unused variable warning in Rust stress test.

+++ bindings/rust/mtls/src/error.rs
@@ -360,7 +360,7 @@ impl Error {
         };
 
         let tls_error: Option<u64> = if c_err.ssl_err != 0 {
-            Some(c_err.ssl_err as u64)
+            Some(c_err.ssl_err)
         } else {
             None
         };
